Handle jQuery.ajax errors properly.
Since the patch step implementation, jQuery's 'error' callback is
wrapped in a promise without being handled properly by the caller,
so the true error callback is never called.

As a result, any failure to load anything prevents the game from
launching, even if the game is perfectly able to cope with the
failure.