Fix serviceworker no longer working when terminated for inactivity.
A service worker may be terminated at any time and recreated without any
notification.  The only persisting state is CacheStorage, which can also
be accessed outside of the service worker.

This patch replaces the global variables from the service worker with
a cache that only contains the path to packed mods.  The complexity
increase should be minimal.