filemanager.js: Explicitly update the service worker on each load.
According to the service worker draft[1], the service worker isn't
guaranteed to be updated when registering it a second time.
Instead, update checks may happen on a fetch event or similar, and are
asynchronous.

This patch looks for an existing service worker registration and
force-update it on each page load.

[1] https://w3c.github.io/ServiceWorker/